apps/backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-pydantic
flask-jwt-extended
python-dotenv
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-google-genai
langchain-groq
google-genai
groq
langchain-chroma
langchain-postgres
langchain-huggingface
pypdf
requests
coverage
psycopg2-binary
faker